refactor: split snail-lower lib.rs into logical modules
Split the large lib.rs file (1455 lines) into smaller, focused modules:

- constants.rs: Public constants and utility functions
- helpers.rs: PyExpr/PyStmt builder functions
- span.rs: Span utility functions
- operators.rs: Operator conversion functions
- awk.rs: AWK-specific lowering logic
- stmt.rs: Statement lowering
- expr.rs: Expression lowering
- program.rs: Top-level program lowering

The lib.rs now serves only as the module declaration and public API
re-export point. All CI checks (fmt, build, clippy, test) pass.
